Fixed bug #1100#.
Deferred use of special options 2.2.3.b and 2.2.3.f, as listed in format
description, until they are added to the system in full.

Deviated from design instructions slightly to improve clarity and conciseness.
Upload deletion now possible.

Made comments window bigger, and now clicking Cancel in the "Edit Comment"
window will send the user back to the "Comments" window instead of just
closing the pop-up altogether, since that behaviour is clearly not appropriate.
Also, changed Cancel button in "Comments" to Done, as that clicking it does
not cancel anything already done.