pkp/pkp-lib#958 Adapt book details depending on pub format settings
This commit adjusts the display of file downloads and publication
format details depending on the number or formats and files. If
a format has only one file, it will display a large download
link with the format named in the link. If a format has multiple
files, it will present an itemized download presentation.

If there is only one publication format, it will list details
such as pub ids, pub dates, etc without naming the format. If
there are multiple publication formats, it will list these under
a heading with the name of the format.