pkp/pkp-lib#2163 Integrate API and Services approach from @kaschioudi
This commit integrates some updates to the REST API infrastructure
from kaschioudi/pkp-lib's rest API branch. These changes include
a new getParameter approach for retrieving params passed to API
endpoints, as well as splitting off a /backend/ API namespace for
specifically serving our own app needs.

It also introduces the services classes introduced in kaschioudi's
services branch. The submission service class is a central place
for managing business logic. At the moment, it provides a common
point for retrieving a list of submissions and deleting
submissions.

These two architectural changes were integrated with the
submissions list panel component that is powered by Vue.js. As a
result, this commit also strips out the large method bundled in
the SubmissionDAO->get() from previous commits. It also strips
out the DBQueryBuilder class that was added, in favor of the
query builder class that kaschioudi has introduced. That builder
is under classes/services/QueryBuilders and is based on the lib
which powers Laravel.