pkp/pkp-lib#2675 Remove search-specific loading indicator in submission lists.
This commit fixes a race condition whereby the search input field was getting
overwritten when the isSearching property was being updated. This caused
some search characters to get dropped if the user typed in the gap between
the isSearching property being changed and the compoonent being re-rendered.
I was unable to determine the precise nature of the problem, so am removing
the search-specific loading indicator in favor of the list component's
generic loading indicator.