Adding manager.navigationMenus.* chains
NavigationMenus is an essential feature, and most of the manager.navigationMenus.* chains are displayed in a first level so they need to be translated.
Website is "Sitio web" (not "página web"). "Website" term consistency need to be reviewed in every translation file, but it will be done for next release (not now).