pkp/pkp-lib#3448 Localize submission titles against submission locale.
This commit introduces a localizeSubmission mixin which will change the
localization fallback order to include the submission locale. It also
fixes some issues with the existing localize mixin to make it behave
like the PHP getLocalizedData methods. Finally, it fixes an error where
empty locale strings were likely to be printed unnecessarily.