pkp/pkp-lib#3130 Anonymize reviewer details in submission logs.
If a user with access to submission logs is also assigned as an author,
this commit will hide reviewer names for blind and double-blind reviews.

To do this, I had to add a new param to the event logs for these events,
so pre-existing open reviews will be anonymized. Only new open reviews
will not be anonymized.