Display journal descriptions without adding <br>
Journal descriptions are edited as rich text, meaning that they contain HTML by default and with TinyMCE as the editor, content will be wrapped in at least `<p>` tags.  Prior to this change, the use of `|nl2br` added a `<br>` tag after every newline, resulting in extra markup/spacing as a description with multiple paragraphs has its `<p>` tags on separate lines:

```html
<p>...</p><br>
<p>...</p><br>
<p>...</p>
```

This change outputs the description verbatim on the page, without conversion.