Update OMP README
A user pointed out that step 4 in the README links to the old PKP documentation wiki, which is, of course, broken. I think this entire step about patches is no longer needed, now that most people install through Git. Can someone confirm this and review the README file to see if anything else needs to be updated?